requirements.txt
dgl-cu102==0.4.3
numpy
tqdm
nltk
prettytable
torch>=1.3.0
zipfile
requests
json
